added fund and stock amount checks

+++ stock/views.py
@@ -86,20 +86,22 @@ def stock_buy(request, stock_ticker, *args, **kwargs):
             tradingID = request.POST.get('selectedAccount')
             shares = form.save(commit=False)
             quantity = shares.shares_amount
-            if Shares.objects.filter(tradingID=tradingID, stockID=stock).exists():
-                shares = Shares.objects.get(tradingID=tradingID, stockID=stock)
-                shares.shares_amount += quantity
-            else:
-                tradingAccount = Trading_Account.objects.filter(pk=tradingID)
-                shares.tradingID= tradingAccount[0]
-                shares.stockID = stock
-            stock.stock_sold += quantity
-            user.userfund.fund-= stock.stock_price * quantity
-            shares.save()
-            stock.save()
-            user.userfund.save()
-            form = SharesForm()
-            return redirect('/stock/buy/'+stock_ticker)
+            if (user.userfund.fund - stock.stock_price * quantity) >= 0:
+                if stock_available - quantity >= 0:
+                    if Shares.objects.filter(tradingID=tradingID, stockID=stock).exists():
+                        shares = Shares.objects.get(tradingID=tradingID, stockID=stock)
+                        shares.shares_amount += quantity
+                    else:
+                        tradingAccount = Trading_Account.objects.filter(pk=tradingID)
+                        shares.tradingID= tradingAccount[0]
+                        shares.stockID = stock
+                    stock.stock_sold += quantity
+                    user.userfund.fund-= stock.stock_price * quantity
+                    shares.save()
+                    stock.save()
+                    user.userfund.save()
+                    form = SharesForm()
+                    return redirect('/stock/buy/'+stock_ticker)
     context = {
         'stock_ticker': stock_ticker,
         'stock_name': stock.stock_name,
@@ -116,23 +118,29 @@ def stock_sell(request, id, stock_ticker, *args, **kwargs):
         stock = Stock.objects.get(stock_ticker=stock_ticker)
     except Stock.DoesNotExist:
         raise Http404
+
     tradingID  = Trading_Account.objects.get(id=id)
-    shares = Shares.objects.get(tradingID=tradingID, stockID=stock)
+    try:
+        shares = Shares.objects.get(tradingID=tradingID, stockID=stock)
+    except Shares.DoesNotExist:
+        raise Http404
+
     user= request.user
 
     form = SharesForm(request.POST or None)
     if request.method == 'POST':
         if form.is_valid():
             quantity = form.cleaned_data['shares_amount']
-            stock.stock_sold -= quantity
-            shares.shares_amount -= quantity
-            user.userfund.fund += stock.stock_price * quantity
-            stock.save()
-            shares.save()
-            user.userfund.save()
-            if shares.shares_amount == 0:
-                shares.delete()
-                return redirect('/trading/')
+            if (shares.shares_amount - quantity) >= 0:
+                stock.stock_sold -= quantity
+                shares.shares_amount -= quantity
+                user.userfund.fund += stock.stock_price * quantity
+                stock.save()
+                shares.save()
+                user.userfund.save()
+                if shares.shares_amount == 0:
+                    shares.delete()
+                    return redirect('/trading/')
             form = SharesForm()
     context = {
     'stock_name': stock.stock_name,
